Add mandatory configuration of an SSH key
Make it so that the SSH private key used for connecting to the runner
instance must be configured as an Ansible variable. This way it will
always be stored in a Vault file under group_vars and the default
identity from the machine running Ansible is not used. This means that
anyone with the Vault key for decrypting the data in the Vault files
will be able to access the runner VM.